Update unit tests and make Google auth credentials optional
